First commit of restructuredText (rst) documentation. Translated some of the existing documentation to rst and deleted these from SVN
RST allows to work on text files and generate beautiful documentation, online (html) as well as offline (PDF via LaTeX). The main advantage is that the documentation is completely done in text files. These allow to be easily be versioned in SVN, as opposed to Word or Writer documents. RST also allows translation of documentation, which is a neat feature.